Implement aggregate sensors for dynamic beamformer config
Implement <beam>.quantiser-gain, .weight and .delay by taking the latest
valid value from the individual engine sensors.

The number of aggregate sensors implemented in generator.py was getting
excessive, so I moved them to a new module (which could benefit from
some testing).

Relates to NGC-447.
